Fixed all test outputs to go into data/test_output, test inputs in da…ta/test_inputs, all of our input files in data/input and our Itinerary and Destination jsons output in data/output. Waiting to do git ignore when I understand more. Things that need to be ignored is all .json and .svg generated files in test_output and output directories.
DO NOT MERGE

There is one outstanding issue which I do not understand. JSONWriterTest is failing at line 56 when comparing JSONWriterTest.json and Sample.json. If you change it to where it is looking to see if the files are not equal, it runs fine... meaning that somehow JSONWriterTest.json and Sample.json are not the exactly the same, even though when you view both of them they are exactly the same. I have tried everything I could think of... using a diff between the files returns nothing in terminal, so I have no idea.