Fixes a bug in web app when adding elements from the search table to the plan table.  The bug occurred when trying to add individual elements unless the number of elements you are adding is greater than the number of elements currently in the plan table.
This bug occurred due to the way the insert function handled the data.  The search table was using a hidden column for indexes which would label all selected items with indexes and use those for the selected item keys.  But, it always started with 0 with the first selected item, and that was causing it to always check the this.state.currentTrip[0]['name'] and locs[0]['name'] resulting in it always thinking that any new selections were dups unless more than the number of items in currentTrip were selected.

For example:  On a fresh page, denver is searched.  One element is selected and the add button is clicked, which will then add that element to the trip table.  Selecting a different element and trying to add would result in that element not being added because index was 0 and the element at index 0 in currentTrip existed... so the code that it was a duplicate entry.  If you instead selected two elements, then we now have indexes 0 and 1; currentTrip is currently length 1, so index 1 does not exist so no dupe and that element gets added, but not element 0.